[BOOK-145] chore: 도서 검색, 도서 등록 API AuthService로 이동
